move code into lib/ansible_viz/
