+++ lib/ansible_viz/grapher.rb
@@ -2,8 +2,8 @@
 # vim: set ts=2 sw=2 tw=100:
 
 require 'rubygems'
-require './graphviz'
-require './styler'
+require 'ansible_viz/graphviz'
+require 'ansible_viz/styler'
 require 'pp'
