+++ lib/ansible_viz/loader.rb
@@ -8,6 +8,7 @@ require 'fileutils'
 require 'ostruct'
 require 'pp'
 
+require 'ansible_viz/utils'
 
 def thing(parent, type, name, path, extra = {})
   human_path = path.sub /^#{ENV['HOME']}/, '~'
