+++ lib/ansible_viz/scoper.rb
