+++ test/varfinder.rb
@@ -4,9 +4,9 @@
 require 'minitest/autorun'
 require 'ostruct'
 
-require './loader'
-require './postprocessor'
-require './scoper'
+require 'ansible_viz/loader'
+require 'ansible_viz/postprocessor'
+require 'ansible_viz/scoper'
 
 
 class TC_FindVars < Minitest::Test
